+++ CHANGELOG.md
@@ -4,11 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2020-12-05
+### Added
+- Link to product attachment
+- Webpack compilation (node v10.23.0)
+
+### Fixed
+- JS error when fast click to plugin on load
+- Access to content in the iframe when not connected
+
 ## [1.0.0] - 2020-11-28
 ### Added
 - Multilanguages
-- Multistore 
+- Multistore
 - Support Prestashop 1.7.1 to 1.7.x
 
 ### Changed
 - ADD link to product / category product / cms / category cms / supplier / manufacturer / others
+